Replaced the Double -> Int.max "dance" with just using UInt64. Thanks for the hint, alexdeem.

+++ Sources/xcresultparser/Models/Coverage/Subrange.swift
@@ -9,6 +9,6 @@ import Foundation
 // Subrange information struct
 struct Subrange: Decodable {
     let column: Int
-    let executionCount: Double
+    let executionCount: UInt64
     let length: Int
 }
